Task 4.2.4: Change the return type of the DistanceCalculation function

+++ app/src/tracker.cpp
@@ -29,5 +29,5 @@ vector<int> Tracker::Tracking(const vector<Rect>& Boxes) {
  * 
  * @return vector<float> 
  */
-int Tracker::DistanceCalculation( int x1, int x2, int y1, int y2) { 
-        return 0; }
+float Tracker::DistanceCalculation( int x1, int x2, int y1, int y2) { 
+        return 0.0; }
